Handle custom errors and change file names
