+++ src/common/helpers/auth.helper.ts
@@ -1,31 +1,50 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import * as jwt from "jsonwebtoken";
 import logger from "../logger/logger";
-import APIRequestError from "../utils/customErrors/apiError";
 
-class Token {
-  createToken = (query: {
-    username: string;
-    passwordHash: string;
-    userId: string;
-  }): string => {
+class AuthHelper {
+  public createAccessToken = (userId: string): string => {
     try {
+      logger.info("get access token", {
+        userId,
+        __filename,
+        functionName: "getAccessToken",
+      });
       const secretKey: string = process.env.SECRET_KEY as string;
-      const token: string = jwt.sign(query, secretKey, { expiresIn: "30d" });
-      logger.info({
-        message: "access token created",
-        data: query,
+      const token = jwt.sign({ id: userId }, secretKey, {
+        expiresIn: "30d",
+      });
+      logger.info("access token created succesfully", {
+        userId,
+        __filename,
+        functionName: "getAccessToken",
       });
       return token;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } catch (err: any) {
-      logger.error({
-        message: `Error in generating token`,
-        error: err.message,
+    } catch (error) {
+      logger.info("token could not be created");
+      throw error;
+    }
+  };
+  public verifyAccessToken = (token: string): any => {
+    try {
+      logger.info("verify access token", {
         __filename,
+        functionName: "verifyAccessToken",
       });
-      throw new APIRequestError("Cannot generate token", query, err);
+      const secretKey: string = process.env.SECRET_KEY as string;
+      const payload = jwt.verify(token, secretKey) as jwt.JwtPayload;
+
+      logger.info(
+        "access token verified successfully",
+        { id: payload.id },
+        __filename,
+        "verifyAccessToken"
+      );
+      return payload;
+    } catch (error) {
+      logger.info("incorrect access token");
+      throw error;
     }
   };
 }
-
-export const token = new Token();
+export const authHelper = new AuthHelper();
