--- src/common/helpers/token.helper.ts
@@ -1,51 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import * as jwt from "jsonwebtoken";
-import { AccessTokenPayload } from "../../data/payloads/access-token.payload";
-import logger from "../logger/logger";
-
-class TokenHelper {
-  public getAccessToken = (accessTokenPayload: AccessTokenPayload): string => {
-    try {
-      logger.info("get access token", {
-        accessTokenPayload,
-        __filename,
-        functionName: "getAccessToken",
-      });
-      const secretKey: string = process.env.SECRET_KEY as string;
-      const token = jwt.sign({ id: accessTokenPayload.id }, secretKey, {
-        expiresIn: "30d",
-      });
-      logger.info("access token created succesfully", {
-        accessTokenPayload,
-        __filename,
-        functionName: "getAccessToken",
-      });
-      return token;
-    } catch (error) {
-      logger.info("token could not be created");
-      throw error;
-    }
-  };
-  public verifyAccessToken = (token: string): any => {
-    try {
-      logger.info("verify access token", {
-        __filename,
-        functionName: "verifyAccessToken",
-      });
-      const secretKey: string = process.env.SECRET_KEY as string;
-      const payload = jwt.verify(token, secretKey) as jwt.JwtPayload;
-
-      logger.info(
-        "access token verified successfully",
-        { id: payload.id },
-        __filename,
-        "verifyAccessToken"
-      );
-      return payload;
-    } catch (error) {
-      logger.info("incorrect access token");
-      throw error;
-    }
-  };
-}
-export const tokenHelper = new TokenHelper();
