+++ src/index.ts
@@ -3,9 +3,9 @@ import { heartbeatController } from "./controllers/heartbeat.controller";
 import { loginRouter } from "./routes/login.route";
 import bodyParser from "body-parser";
 import * as error from "./middleware/error.middleware";
-import { redisHelper } from "./common/helpers/redis.helper";
+import { redisHelper } from "./storage/redis.helper";
+import { PORT } from "./common/constants";
 
-const PORT = 3000;
 const app: Express = express();
 
 (async () => {
